refactor: factored out ParameterPanel
